Add loading message
While testing on production it was found that it was taking longer than usual so clicked the button multiple times which caused multiple responses/facts to appear so a 'Loading...' message has been set untill the fact appears. Also if there is some error in fetching the fact then 'Failed to Get Fact' message will appear.